@@ -32,34 +32,44 @@
 #
 # Contributor(s): Pavel Císař (original code)
 #                 ______________________________________
-# pylint: disable=C0302, W0212, R0902, R0912,R0913, R0914, R0915, R0904, R0903, C0301, W0703
 
-"""firebird.lib.trace - Module for parsing Firebird trace & audit protocol
+"""firebird.lib.trace - Module for parsing Firebird textual trace & audit logs.
 
+This module provides classes and a parser (`TraceParser`) for processing the
+standard text output generated by a Firebird trace session (configured via
+`fbtrace.conf` or similar).
 
+The parser analyzes the log line by line or block by block, identifying different
+trace events (like database attaches, transaction starts, statement executions)
+and associated information blocks (like connection details, SQL text, parameters).
+It yields instances of `.TraceEvent` subclasses (e.g., `.EventAttach`,
+`.EventStatementStart`) and `.TraceInfo` subclasses (e.g., `.AttachmentInfo`,
+`.SQLInfo`) representing the parsed data.
 """
 
 from __future__ import annotations
-from typing import List, Tuple, Dict, Set, Iterable, Any, Optional, Union
-from sys import intern
+
+import collections
 import datetime
 import decimal
-import collections
-from enum import Enum, IntEnum, auto
+from collections.abc import Iterable
 from dataclasses import dataclass
-from firebird.base.types import Error, STOP, Sentinel
+from enum import Enum, IntEnum, auto
+from sys import intern
+from typing import Any
+
+from firebird.base.types import STOP, Error
+
 
 class Status(Enum):
-    """Trace event status codes.
-    """
+    """Represents the outcome status of a traced operation."""
     OK = ' '
     FAILED = 'F'
     UNAUTHORIZED = 'U'
     UNKNOWN = '?'
 
 class Event(IntEnum):
-    """Trace event codes.
-    """
+    """Enumerates the different types of events captured in a trace log."""
     UNKNOWN = auto()
     TRACE_INIT = auto()
     TRACE_SUSPENDED = auto()
@@ -100,12 +110,12 @@ class Event(IntEnum):
     EXECUTE_DYN = auto()
 
 class TraceInfo:
-    """Base class for trace info blocks.
-    """
+    """Base class for informational blocks extracted from the trace log
+    (e.g., attachment details, SQL text)."""
 
 class TraceEvent:
-    """Base class for trace events.
-    """
+    """Base class for specific events recorded in the trace log
+    (e.g., attach database, start transaction)."""
 
 @dataclass(frozen=True)
 class AttachmentInfo(TraceInfo):
@@ -140,8 +150,8 @@ class TransactionInfo(TraceInfo):
     transaction_id: int
     #: Initial transaction ID (for retained ones)
     initial_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
 
 @dataclass(frozen=True)
 class ServiceInfo(TraceInfo):
@@ -177,8 +187,8 @@ class ParamSet(TraceInfo):
     """
     #: Parameter set ID
     par_id: int
-    #: List of parameters (name, value pairs)
-    params: List[Tuple[str, Any]]
+    #: list of parameters (name, value pairs)
+    params: list[tuple[str, Any]]
 
 @dataclass(frozen=True)
 class AccessStats(TraceInfo):
@@ -364,8 +374,8 @@ class EventTransactionStart(TraceEvent):
     attachment_id: int
     #: Transaction ID
     transaction_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
 
 @dataclass(frozen=True)
 class EventCommit(TraceEvent):
@@ -380,8 +390,8 @@ class EventCommit(TraceEvent):
     attachment_id: int
     #: Transaction ID
     transaction_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
     #: Execution time in ms
     run_time: int
     #: Number of page reads
@@ -406,8 +416,8 @@ class EventRollback(TraceEvent):
     attachment_id: int
     #: Transaction ID
     transaction_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
     #: Execution time in ms
     run_time: int
     #: Number of page reads
@@ -432,8 +442,8 @@ class EventCommitRetaining(TraceEvent):
     attachment_id: int
     #: Transaction ID
     transaction_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
     #: New transaction number
     new_transaction_id: int
     #: Execution time in ms
@@ -460,8 +470,8 @@ class EventRollbackRetaining(TraceEvent):
     attachment_id: int
     #: Transaction ID
     transaction_id: int
-    #: List of transaction options
-    options: List[str]
+    #: list of transaction options
+    options: list[str]
     #: New transaction number
     new_transaction_id: int
     #: Execution time in ms
@@ -547,8 +557,8 @@ class EventStatementFinish(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 
 @dataclass(frozen=True)
 class EventFreeStatement(TraceEvent):
@@ -632,8 +642,8 @@ class EventTriggerFinish(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 
 #
 @dataclass(frozen=True)
@@ -683,8 +693,8 @@ class EventProcedureFinish(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 #
 @dataclass(frozen=True)
 class EventFunctionStart(TraceEvent):
@@ -722,7 +732,7 @@ class EventFunctionFinish(TraceEvent):
     #: Param set ID (ParamSet)
     param_id: int
     #: Return value
-    returns: Tuple[str, Any]
+    returns: tuple[str, Any]
     #: Execution time in ms
     run_time: int
     #: Number of page reads
@@ -733,8 +743,8 @@ class EventFunctionFinish(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 #
 @dataclass(frozen=True)
 class EventServiceAttach(TraceEvent):
@@ -773,8 +783,8 @@ class EventServiceStart(TraceEvent):
     service_id: int
     #: Action performed by service
     action: str
-    #: List of action parameters
-    parameters: List[str]
+    #: list of action parameters
+    parameters: list[str]
 
 @dataclass(frozen=True)
 class EventServiceQuery(TraceEvent):
@@ -789,15 +799,15 @@ class EventServiceQuery(TraceEvent):
     service_id: int
     #: Action performed by service
     action: str
-    #: List of sent items
-    sent: List[str]
-    #: List of received items
-    received: List[str]
+    #: list of sent items
+    sent: list[str]
+    #: list of received items
+    received: list[str]
 
 #
 @dataclass(frozen=True)
 class EventSetContext(TraceEvent):
-    "Set context variable trace event"
+    "set context variable trace event"
     #: Trace event ID
     event_id: int
     #: Timestamp when the event occurred
@@ -826,7 +836,7 @@ class EventError(TraceEvent):
     #: Place where error occured
     place: str
     #: Error details
-    details: List[str]
+    details: list[str]
 
 @dataclass(frozen=True)
 class EventWarning(TraceEvent):
@@ -840,7 +850,7 @@ class EventWarning(TraceEvent):
     #: Place where warning occured
     place: str
     #: Warning details
-    details: List[str]
+    details: list[str]
 
 @dataclass(frozen=True)
 class EventServiceError(TraceEvent):
@@ -854,7 +864,7 @@ class EventServiceError(TraceEvent):
     #: Place where error occured
     place: str
     #: Error details
-    details: List[str]
+    details: list[str]
 
 @dataclass(frozen=True)
 class EventServiceWarning(TraceEvent):
@@ -868,7 +878,7 @@ class EventServiceWarning(TraceEvent):
     #: Place where warning occured
     place: str
     #: Warning details
-    details: List[str]
+    details: list[str]
 
 #
 @dataclass(frozen=True)
@@ -908,8 +918,8 @@ class EventSweepProgress(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 
 @dataclass(frozen=True)
 class EventSweepFinish(TraceEvent):
@@ -938,8 +948,8 @@ class EventSweepFinish(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 
 @dataclass(frozen=True)
 class EventSweepFailed(TraceEvent):
@@ -997,8 +1007,8 @@ class EventBLRExecute(TraceEvent):
     fetches: int
     #: Number of pages with changes pending
     marks: int
-    #: List with table access statistics
-    access: List[AccessStats]
+    #: list with table access statistics
+    access: list[AccessStats]
 
 @dataclass(frozen=True)
 class EventDYNExecute(TraceEvent):
@@ -1029,25 +1039,46 @@ class EventUnknown(TraceEvent):
     data: str
 
 def safe_int(str_value: str, base: int=10):
-    """Always returns integer value from string/None argument. Returns 0 if argument is None.
+    """Safely converts a string to an integer, returning 0 for None or empty string.
+
+    Arguments:
+        str_value: The string to convert.
+        base: The base for the conversion (default is 10).
+
+    Returns:
+        The integer value, or 0 if `str_value` is None or empty/whitespace.
     """
     if str_value:
         return int(str_value, base)
     return 0
 
 class TraceParser:
-    """Parser for standard textual trace log. Produces dataclasses describing individual
-    trace log entries/events.
+    """Parses the textual output of a Firebird trace session.
+
+    This class processes lines from a trace log, identifying individual events
+    and associated information blocks. It maintains internal state to track
+    database attachments, transactions, services, unique SQL statements, and
+    parameter sets, assigning unique IDs where appropriate.
+
+    It can be used in two ways:
+
+    1.  Iteratively via `parse(lines)`: Processes an entire iterable of log lines
+        and yields parsed `.TraceEvent` and `.TraceInfo` objects.
+    2.  Incrementally via `push(line)`: Processes one line at a time, returning
+        a list of parsed objects when a complete event block is detected.
+
+    Parsed `TraceInfo` objects related to an event are typically yielded/returned
+    *before* the corresponding `.TraceEvent` object.
     """
     def __init__(self):
-        #: Set of attachment ids that were already processed
-        self.seen_attachments: Set[int] = set()
-        #: Set of transaction ids that were already processed
-        self.seen_transactions: Set[int] = set()
-        #: Set of service ids that were already processed
+        #: set of attachment ids that were already processed
+        self.seen_attachments: set[int] = set()
+        #: set of transaction ids that were already processed
+        self.seen_transactions: set[int] = set()
+        #: set of service ids that were already processed
         self.seen_services: set[int] = set()
         #: Dictionary that maps (sql_cmd, plan) keys to internal ids
-        self.sqlinfo_map: Dict[Tuple[str, str], int]= {}
+        self.sqlinfo_map: dict[tuple[str, str], int]= {}
         #: Dictionary that maps parameters (statement or procedure) keys to internal ids
         self.param_map = {}
         #: Sequence id that would be assigned to next parsed event (starts with 1)
@@ -1063,10 +1094,10 @@ class TraceParser:
         self.has_statement_free: bool = True
         #
         self.__infos: collections.deque = collections.deque()
-        self.__pushed: List[str] = []
+        self.__pushed: list[str] = []
         self.__current_block: collections.deque = collections.deque()
         self.__last_timestamp: datetime.datetime = None
-        self.__event_values: Dict[str, Any] = {}
+        self.__event_values: dict[str, Any] = {}
         self.__parse_map = {Event.TRACE_INIT: self.__parser_trace_init,
                             Event.TRACE_FINI: self.__parser_trace_finish,
                             Event.START_TRANSACTION: self.__parser_start_transaction,
@@ -1107,7 +1138,7 @@ class TraceParser:
     def _is_entry_header(self, line: str) -> bool:
         items = line.split()
         try:
-            datetime.datetime.strptime(items[0], '%Y-%m-%dT%H:%M:%S.%f')
+            datetime.datetime.strptime(items[0], '%Y-%m-%dT%H:%M:%S.%f') # noqa: DTZ007
             return True
         except Exception:
             return False
@@ -1128,7 +1159,7 @@ class TraceParser:
     def _iter_trace_blocks(self, ilines):
         lines = []
         for line in ilines:
-            line = line.strip()
+            line = line.strip() # noqa: PLW2901
             if line:
                 if not lines:
                     if self._is_entry_header(line):
@@ -1152,7 +1183,7 @@ class TraceParser:
         if items[2] == 'Unknown':
             return Event.UNKNOWN
         raise Error(f'Unrecognized event header: "{line}"')
-    def _parse_attachment_info(self, values: Dict[str, Any], check: bool=True) -> None:
+    def _parse_attachment_info(self, values: dict[str, Any], *, check: bool=True) -> None:
         database, _, attachment = self.__current_block.popleft().partition(' (')
         values['database'] = intern(database)
         attachment_id, user_role, charset, protocol_address = attachment.strip('()').split(',')
@@ -1197,7 +1228,7 @@ class TraceParser:
         if check and values['attachment_id'] not in self.seen_attachments:
             self.__infos.append(AttachmentInfo(**values))
         self.seen_attachments.add(values['attachment_id'])
-    def _parse_transaction_info(self, values: Dict[str, Any], check: bool=True) -> None:
+    def _parse_transaction_info(self, values: dict[str, Any], *, check: bool=True) -> None:
         # Transaction parameters
         items = self.__current_block.popleft().strip('\t ()').split(',')
         if len(items) == 2:
@@ -1222,7 +1253,7 @@ class TraceParser:
         self.__event_values['marks'] = None
         if self.__current_block:
             for value in self.__current_block.popleft().split(','):
-                value, val_type = value.split()
+                value, val_type = value.split() # noqa: PLW2901
                 if 'ms' in val_type:
                     self.__event_values['run_time'] = int(value)
                 elif 'read' in val_type:
@@ -1313,7 +1344,7 @@ class TraceParser:
             if line:
                 self.__current_block.appendleft(line)
             self.__event_values['plan'] = intern('\n'.join(plan))
-    def _parse_value_spec(self, param_def: str) -> Tuple[str, Any]:
+    def _parse_value_spec(self, param_def: str) -> tuple[str, Any]:
         param_type, param_value = param_def.split(',', 1)
         param_type = intern(param_type)
         param_value = param_value.strip(' "')
@@ -1322,22 +1353,22 @@ class TraceParser:
         elif param_type in ('smallint', 'integer', 'bigint'):
             param_value = int(param_value)
         elif param_type == 'timestamp':
-            param_value = datetime.datetime.strptime(param_value, '%Y-%m-%dT%H:%M:%S.%f')
+            param_value = datetime.datetime.strptime(param_value, '%Y-%m-%dT%H:%M:%S.%f') # noqa: DTZ007
         elif param_type == 'date':
-            param_value = datetime.datetime.strptime(param_value, '%Y-%m-%d')
+            param_value = datetime.datetime.strptime(param_value, '%Y-%m-%d') # noqa: DTZ007
         elif param_type == 'time':
-            param_value = datetime.datetime.strptime(param_value, '%H:%M:%S.%f')
+            param_value = datetime.datetime.strptime(param_value, '%H:%M:%S.%f') # noqa: DTZ007
         elif param_type in ('float', 'double precision'):
             param_value = decimal.Decimal(param_value)
         return (param_type, param_value,)
-    def _parse_parameters_block(self) -> List[Tuple[str, Any]]:
+    def _parse_parameters_block(self) -> list[tuple[str, Any]]:
         parameters = []
         while self.__current_block and self.__current_block[0].startswith('param'):
             line = self.__current_block.popleft()
             _, param_def = line.split(' = ')
             parameters.append(self._parse_value_spec(param_def))
         return parameters
-    def _parse_parameters(self, for_procedure: bool=False) -> None:
+    def _parse_parameters(self, *, for_procedure: bool=False) -> None:
         parameters = self._parse_parameters_block()
         while self.__current_block and self.__current_block[0].endswith('more arguments skipped...'):
             self.__current_block.popleft()
@@ -1487,7 +1518,7 @@ class TraceParser:
     def __parse_trace_header(self) -> None:
         line = self.__current_block.popleft()
         items = line.split()
-        self.__last_timestamp = datetime.datetime.strptime(items[0], '%Y-%m-%dT%H:%M:%S.%f')
+        self.__last_timestamp = datetime.datetime.strptime(items[0], '%Y-%m-%dT%H:%M:%S.%f') # noqa: DTZ007
         if (len(items) == 3) or (items[2] in ('ERROR', 'WARNING')):
             self.__event_values['status'] = Status.OK
         else:
@@ -1781,7 +1812,7 @@ class TraceParser:
         self.__event_values['value'] = value.strip(' "')
         del self.__event_values['status']
         return EventSetContext(**self.__event_values)
-    def __parser_error(self) -> Union[EventServiceError, EventError]:
+    def __parser_error(self) -> EventServiceError | EventError:
         self.__event_values['place'] = self.__current_block[0].split(' AT ')[1]
         self.__parse_trace_header()
         att_values = {}
@@ -1798,7 +1829,7 @@ class TraceParser:
         self.__event_values['details'] = details
         del self.__event_values['status']
         return event_class(**self.__event_values)
-    def __parser_warning(self) -> Union[EventServiceWarning, EventWarning]:
+    def __parser_warning(self) -> EventServiceWarning | EventWarning:
         self.__event_values['place'] = self.__current_block[0].split(' AT ')[1]
         self.__parse_trace_header()
         att_values = {}
@@ -1876,7 +1907,7 @@ class TraceParser:
         del self.__event_values['status']
         self.__event_values['data'] = '\n'.join(self.__current_block)
         return EventUnknown(**self.__event_values)
-    def retrieve_info(self) -> List[TraceInfo]:
+    def retrieve_info(self) -> list[TraceInfo]:
         """Returns list of `.TraceInfo` instances produced by last `.parse_event()` call.
 
         The list could be empty.
@@ -1889,11 +1920,11 @@ class TraceParser:
         result = self.__infos.copy()
         self.__infos.clear()
         return result
-    def parse_event(self, trace_block: List[str]) -> TraceEvent:
+    def parse_event(self, trace_block: list[str]) -> TraceEvent:
         """Parse single trace event.
 
         Arguments:
-            trace_block: List with trace entry lines for single trace event.
+            trace_block: list with trace entry lines for single trace event.
         """
         self.__current_block.clear()
         self.__current_block.extend(trace_block)
@@ -1918,7 +1949,7 @@ class TraceParser:
             while len(self.__infos) > 0:
                 yield self.__infos.popleft()
             yield rec
-    def push(self, line: Union[str, Sentinel]) -> Optional[List[Union[TraceEvent, TraceInfo]]]:
+    def push(self, line: str | STOP) -> list[TraceEvent| TraceInfo] | None:
         """Push parser.
 
         Arguments:
