@@ -1,9 +1,11 @@
-#coding:utf-8
+# SPDX-FileCopyrightText: 2020-present The Firebird Projects <www.firebirdsql.org>
+#
+# SPDX-License-Identifier: MIT
 #
 # PROGRAM/MODULE: firebird-lib
-# FILE:           test_log.py
-# DESCRIPTION:    Unit tests for firebird.lib.log
-# CREATED:        8.10.2020
+# FILE:           tests/test_log.py
+# DESCRIPTION:    Tests for firebird.lib.log module
+# CREATED:        25.4.2025
 #
 # The contents of this file are subject to the MIT License
 #
@@ -29,176 +31,98 @@
 # All Rights Reserved.
 #
 # Contributor(s): Pavel Císař (original code)
-#                 ______________________________________
 
-"""firebird-lib - Unit tests for firebird.lib.log
-
-
+"""firebird-lib - Tests for firebird.lib.log module
 """
 
-import unittest
-import sys, os
+import pytest
 from collections.abc import Sized, MutableSequence, Mapping
 from re import finditer
 from io import StringIO
-from firebird.driver import *
 from firebird.lib.log import *
 
+# --- Constants ---
 FB30 = '3.0'
 FB40 = '4.0'
 FB50 = '5.0'
 
-if driver_config.get_server('local') is None:
-    # Register Firebird server
-    srv_cfg = """[local]
-    host = localhost
-    user = SYSDBA
-    password = masterkey
-    """
-    driver_config.register_server('local', srv_cfg)
-
-# Register database
-if driver_config.get_database('fbtest') is None:
-    db_cfg = """[fbtest]
-    server = local
-    database = fbtest3.fdb
-    protocol = inet
-    charset = utf8
-    """
-    driver_config.register_database('fbtest', db_cfg)
+# --- Helper Functions ---
 
 def linesplit_iter(string):
-    return (m.group(2) for m in finditer('((.*)\n|(.+)$)', string))
+    """Iterates over lines in a string, handling different line endings."""
+    # Add handling for potential None groups if string ends exactly with \n
+    return (m.group(2) or m.group(3) or ''
+            for m in finditer('((.*)\n|(.+)$)', string))
 
 def iter_obj_properties(obj):
-    """Iterator function.
-
-    Args:
-        obj (class): Class object.
-
-    Yields:
-        `name', 'property` pairs for all properties in class.
-"""
+    """Iterator function for object properties."""
     for varname in dir(obj):
         if hasattr(type(obj), varname) and isinstance(getattr(type(obj), varname), property):
             yield varname
 
 def iter_obj_variables(obj):
-    """Iterator function.
-
-    Args:
-        obj (class): Class object.
-
-    Yields:
-        Names of all non-callable attributes in class.
-"""
+    """Iterator function for object variables (non-callable, non-private)."""
     for varname in vars(obj):
         value = getattr(obj, varname)
         if not callable(value) and not varname.startswith('_'):
             yield varname
 
 def get_object_data(obj, skip=[]):
+    """Extracts attribute and property data from an object into a dictionary."""
+    data = {}
     def add(item):
         if item not in skip:
             value = getattr(obj, item)
+            # Store length for sized collections/mappings instead of the full object
             if isinstance(value, Sized) and isinstance(value, (MutableSequence, Mapping)):
                 value = len(value)
             data[item] = value
 
-    data = {}
     for item in iter_obj_variables(obj):
         add(item)
     for item in iter_obj_properties(obj):
         add(item)
     return data
 
-class TestBase(unittest.TestCase):
-    def __init__(self, methodName='runTest'):
-        super(TestBase, self).__init__(methodName)
-        self.output = StringIO()
-        self.FBTEST_DB = 'fbtest'
-    def setUp(self):
-        with connect_server('local') as svc:
-            self.version = svc.info.version
-        if self.version.startswith('3.0'):
-            self.FBTEST_DB = 'fbtest30.fdb'
-            self.version = FB30
-        elif self.version.startswith('4.0'):
-            self.FBTEST_DB = 'fbtest40.fdb'
-            self.version = FB40
-        elif self.version.startswith('5.0'):
-            self.FBTEST_DB = 'fbtest50.fdb'
-            self.version = FB50
-        else:
-            raise Exception("Unsupported Firebird version (%s)" % self.version)
-        #
-        self.cwd = os.getcwd()
-        self.dbpath = self.cwd if os.path.split(self.cwd)[1] == 'test' \
-            else os.path.join(self.cwd, 'test')
-        self.dbfile = os.path.join(self.dbpath, self.FBTEST_DB)
-        driver_config.get_database('fbtest').database.value = self.dbfile
-    def clear_output(self):
-        self.output.close()
-        self.output = StringIO()
-    def show_output(self):
-        sys.stdout.write(self.output.getvalue())
-        sys.stdout.flush()
-    def printout(self, text='', newline=True, no_rstrip=False):
-        if no_rstrip:
-            self.output.write(text)
-        else:
-            self.output.write(text.rstrip())
-        if newline:
-            self.output.write('\n')
-        self.output.flush()
-    def printData(self, cur, print_header=True):
-        """Print data from open cursor to stdout."""
-        if print_header:
-            # Print a header.
-            line = []
-            for fieldDesc in cur.description:
-                line.append(fieldDesc[DESCRIPTION_NAME].ljust(fieldDesc[DESCRIPTION_DISPLAY_SIZE]))
-            self.printout(' '.join(line))
-            line = []
-            for fieldDesc in cur.description:
-                line.append("-" * max((len(fieldDesc[DESCRIPTION_NAME]), fieldDesc[DESCRIPTION_DISPLAY_SIZE])))
-            self.printout(' '.join(line))
-        # For each row, print the value of each field left-justified within
-        # the maximum possible width of that field.
-        fieldIndices = range(len(cur.description))
-        for row in cur:
-            line = []
-            for fieldIndex in fieldIndices:
-                fieldValue = str(row[fieldIndex])
-                fieldMaxWidth = max((len(cur.description[fieldIndex][DESCRIPTION_NAME]), cur.description[fieldIndex][DESCRIPTION_DISPLAY_SIZE]))
-                line.append(fieldValue.ljust(fieldMaxWidth))
-            self.printout(' '.join(line))
+# --- Test Helper Functions ---
 
-class TestLogParser(TestBase):
-    def setUp(self):
-        super().setUp()
-        self.dbfile = os.path.join(self.dbpath, self.FBTEST_DB)
-        self.maxDiff = None
-    def _check_events(self, log_lines, output):
-        self.output = StringIO()
-        parser = LogParser()
+def _parse_log_lines(log_lines: str) -> str:
+    """Parses log lines using LogParser.parse and returns string representation."""
+    output_io = StringIO()
+    parser = LogParser()
+    # Handle potential StopIteration if log_lines is empty
+    try:
         for obj in parser.parse(linesplit_iter(log_lines)):
-            self.printout(str(obj))
-        self.assertEqual(self.output.getvalue(), output, "PARSE: Parsed events do not match expected ones")
-        self._push_check_events(log_lines, output)
-        self.output.close()
-    def _push_check_events(self, log_lines, output):
-        self.output = StringIO()
-        parser = LogParser()
-        for line in linesplit_iter(log_lines):
-            if event := parser.push(line):
-                self.printout(str(event))
-        if event := parser.push(STOP):
-            self.printout(str(event))
-        self.assertEqual(self.output.getvalue(), output, "PUSH: Parsed events do not match expected ones")
-        self.output.close()
-    def test_01_win_fb2_with_unknown(self):
-        data = """
+            print(str(obj), file=output_io, end='\n') # Ensure newline
+    except StopIteration:
+        pass # No events parsed from empty input
+    return output_io.getvalue()
+
+def _push_log_lines(log_lines: str) -> str:
+    """Parses log lines using LogParser.push and returns string representation."""
+    output_io = StringIO()
+    parser = LogParser()
+    for line in linesplit_iter(log_lines):
+        events = parser.push(line)
+        if events: # push might return a list or a single event
+            if not isinstance(events, list):
+                events = [events]
+            for event in events:
+                print(str(event), file=output_io, end='\n') # Ensure newline
+    # Process any remaining buffered lines
+    final_events = parser.push(STOP)
+    if final_events:
+        if not isinstance(final_events, list):
+            final_events = [final_events]
+        for event in final_events:
+            print(str(event), file=output_io, end='\n') # Ensure newline
+    return output_io.getvalue()
+
+# --- Test Functions ---
+
+def test_01_win_fb2_with_unknown():
+    """Tests parsing a log fragment from Windows FB 2.x with unknown host messages."""
+    data = """
 
 SRVDB1  Tue Apr 04 21:25:40 2017
         INET/inet_error: read errno = 10054
@@ -274,7 +198,7 @@ SRVDB1  Thu Apr 06 12:52:56 2017
 
 
 """
-        output = """LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 25, 40), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'read', 'err_code': 10054})
+    expected_output = """LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 25, 40), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'read', 'err_code': 10054})
 LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 25, 41), level=<Severity.UNKNOWN: 0>, code=0, facility=<Facility.UNKNOWN: 0>, message='Unable to complete network request to host "SRVDB1".\\nError reading data from the connection.', params={})
 LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 25, 42), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'read', 'err_code': 10054})
 LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 25, 43), level=<Severity.UNKNOWN: 0>, code=0, facility=<Facility.UNKNOWN: 0>, message='Unable to complete network request to host "SRVDB1".\\nError reading data from the connection.', params={})
@@ -290,9 +214,16 @@ LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 28, 57),
 LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 4, 21, 28, 58), level=<Severity.UNKNOWN: 0>, code=0, facility=<Facility.UNKNOWN: 0>, message='Unable to complete network request to host "(unknown)".\\nError reading data from the connection.', params={})
 LogMessage(origin='SRVDB1', timestamp=datetime.datetime(2017, 4, 6, 12, 52, 56), level=<Severity.WARNING: 2>, code=73, facility=<Facility.SYSTEM: 1>, message='Shutting down the server with {con_count} active connection(s) to {db_count} database(s), {svc_count} active service(s)', params={'con_count': 1, 'db_count': 1, 'svc_count': 0})
 """
-        self._check_events(data, output)
-    def test_02_lin_fb3(self):
-        data = """
+    # Using strip() to handle potential trailing newline differences
+    parsed_output = _parse_log_lines(data)
+    assert parsed_output.strip() == expected_output.strip(), "PARSE: Parsed events do not match expected ones"
+
+    pushed_output = _push_log_lines(data)
+    assert pushed_output.strip() == expected_output.strip(), "PUSH: Parsed events do not match expected ones"
+
+def test_02_lin_fb3():
+    """Tests parsing a log fragment from Linux FB 3.x with guardian messages."""
+    data = """
 MyServer (Client)	Fri Apr  6 16:35:46 2018
 	INET/inet_error: connect errno = 111
 
@@ -338,7 +269,7 @@ MyServer (Client)	Tue Apr 17 19:42:55 2018
 
 
 """
-        output = """LogMessage(origin='MyServer (Client)', timestamp=datetime.datetime(2018, 4, 6, 16, 35, 46), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'connect', 'err_code': 111})
+    expected_output = """LogMessage(origin='MyServer (Client)', timestamp=datetime.datetime(2018, 4, 6, 16, 35, 46), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'connect', 'err_code': 111})
 LogMessage(origin='MyServer (Client)', timestamp=datetime.datetime(2018, 4, 6, 16, 51, 31), level=<Severity.INFO: 1>, code=151, facility=<Facility.GUARDIAN: 9>, message='{prog_name}: guardian starting {value}', params={'prog_name': '/opt/firebird/bin/fbguard', 'value': '/opt/firebird/bin/fbserver'})
 LogMessage(origin='MyServer (Server)', timestamp=datetime.datetime(2018, 4, 6, 16, 55, 23), level=<Severity.INFO: 1>, code=124, facility=<Facility.SYSTEM: 1>, message='activating shadow file {shadow}', params={'shadow': '/home/db/test_employee.fdb'})
 LogMessage(origin='MyServer (Server)', timestamp=datetime.datetime(2018, 4, 6, 16, 55, 31), level=<Severity.INFO: 1>, code=126, facility=<Facility.SWEEP: 7>, message='Sweep is started by {user}\\nDatabase "{database}"\\nOIT {oit}, OAT {oat}, OST {ost}, Next {next}', params={'user': 'SYSDBA', 'database': '/home/db/test_employee.fdb', 'oit': 1, 'oat': 0, 'ost': 0, 'next': 1})
@@ -348,9 +279,15 @@ LogMessage(origin='MyServer (Client)', timestamp=datetime.datetime(2018, 4, 9, 8
 LogMessage(origin='MyServer (Server)', timestamp=datetime.datetime(2018, 4, 17, 15, 1, 27), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}', params={'error': 'invalid socket in packet_receive', 'err_code': 22})
 LogMessage(origin='MyServer (Client)', timestamp=datetime.datetime(2018, 4, 17, 19, 42, 55), level=<Severity.INFO: 1>, code=162, facility=<Facility.GUARDIAN: 9>, message='{prog_name}: {process_name} normal shutdown.', params={'prog_name': '/opt/firebird/bin/fbguard', 'process_name': '/opt/firebird/bin/fbserver'})
 """
-        self._check_events(data, output)
-    def test_03_lin_fb3(self):
-        data = """
+    parsed_output = _parse_log_lines(data)
+    assert parsed_output.strip() == expected_output.strip(), "PARSE: Parsed events do not match expected ones"
+
+    pushed_output = _push_log_lines(data)
+    assert pushed_output.strip() == expected_output.strip(), "PUSH: Parsed events do not match expected ones"
+
+def test_03_lin_fb3_validation_auth_etc():
+    """Tests parsing a log fragment from Linux FB 3.x with various messages."""
+    data = """
 ultron	Sun Oct 28 15:25:54 2018
 	/opt/firebird/bin/fbguard: guardian starting /opt/firebird/bin/firebird
 
@@ -399,7 +336,7 @@ ultron	Thu Jun 13 07:36:41 2019
 
 
 """
-        output = """LogMessage(origin='ultron', timestamp=datetime.datetime(2018, 10, 28, 15, 25, 54), level=<Severity.INFO: 1>, code=151, facility=<Facility.GUARDIAN: 9>, message='{prog_name}: guardian starting {value}', params={'prog_name': '/opt/firebird/bin/fbguard', 'value': '/opt/firebird/bin/firebird'})
+    expected_output = """LogMessage(origin='ultron', timestamp=datetime.datetime(2018, 10, 28, 15, 25, 54), level=<Severity.INFO: 1>, code=151, facility=<Facility.GUARDIAN: 9>, message='{prog_name}: guardian starting {value}', params={'prog_name': '/opt/firebird/bin/fbguard', 'value': '/opt/firebird/bin/firebird'})
 LogMessage(origin='ultron', timestamp=datetime.datetime(2018, 10, 28, 15, 29, 42), level=<Severity.INFO: 1>, code=157, facility=<Facility.GUARDIAN: 9>, message='{prog_name}: {process_name} terminated', params={'prog_name': '/opt/firebird/bin/fbguard', 'process_name': '/opt/firebird/bin/firebird'})
 LogMessage(origin='ultron', timestamp=datetime.datetime(2018, 10, 31, 13, 47, 44), level=<Severity.ERROR: 3>, code=284, facility=<Facility.SYSTEM: 1>, message='REMOTE INTERFACE/gds__detach: Unsuccesful detach from database.\\nUncommitted work may have been lost.\\n{err_msg}', params={'err_msg': 'Error writing data to the connection.'})
 LogMessage(origin='ultron', timestamp=datetime.datetime(2018, 11, 14, 3, 32, 44), level=<Severity.ERROR: 3>, code=177, facility=<Facility.NET: 10>, message='INET/inet_error: {error} errno = {err_code}, {parameters}', params={'error': 'read', 'err_code': 104, 'parameters': 'client host = Terminal, address = 192.168.1.243/55120, user = frodo'})
@@ -409,4 +346,82 @@ LogMessage(origin='ultron', timestamp=datetime.datetime(2019, 6, 13, 7, 32, 51),
 LogMessage(origin='ultron', timestamp=datetime.datetime(2019, 6, 13, 7, 36, 41), level=<Severity.WARNING: 2>, code=81, facility=<Facility.VALIDATION: 6>, message='Database: {database}\\nWarning: Page {page_num} is an orphan', params={'database': '/usr/local/data/mydb.FDB', 'page_num': 3867207})
 LogMessage(origin='ultron', timestamp=datetime.datetime(2019, 6, 13, 7, 36, 41), level=<Severity.INFO: 1>, code=76, facility=<Facility.VALIDATION: 6>, message='Database: {database}\\nValidation finished: {errors} errors, {warnings} warnings, {fixed} fixed', params={'database': '/usr/local/data/mydb.FDB', 'errors': 0, 'warnings': 663, 'fixed': 663})
 """
-        self._check_events(data, output)
+    parsed_output = _parse_log_lines(data)
+    assert parsed_output.strip() == expected_output.strip(), "PARSE: Parsed events do not match expected ones"
+
+    pushed_output = _push_log_lines(data)
+    assert pushed_output.strip() == expected_output.strip(), "PUSH: Parsed events do not match expected ones"
+
+def test_04_parse_entry_malformed_header():
+    parser = LogParser()
+    malformed_entry = ["MyOrigin Invalid Date String 2023", "  Continuation line"]
+    with pytest.raises(Error, match="Malformed log entry"):
+        parser.parse_entry(malformed_entry)
+
+def test_05_push_malformed_header_error():
+    parser = LogParser()
+    lines = ["MyOrigin Invalid Date String 2023"]
+    # Need to ensure this specific line bypasses push's own date check if possible,
+    # or structure the test data carefully. This might be tricky as push
+    # might treat it as a continuation. A direct parse_entry test is safer.
+    # If push treats it as continuation, test the result after STOP.
+    parser.push(lines[0])
+    with pytest.raises(Error, match="Malformed log entry"):
+        parser.push(STOP) # Error likely occurs here when parse_entry is called
+
+def test_06_parse_empty_input():
+    parser = LogParser()
+    lines = []
+    with pytest.raises(Error, match="Malformed log entry"):
+        list(parser.parse(lines))
+
+def test_07_push_state_transitions():
+    parser = LogParser()
+    # 1. Push header - should return None
+    line1 = "Origin1 Tue Apr 04 21:25:40 2017"
+    assert parser.push(line1) is None
+    assert len(parser._LogParser__buffer) == 1 # Access internal for verification
+
+    # 2. Push continuation - should return None
+    line2 = "  Continuation line 1"
+    assert parser.push(line2) is None
+    assert len(parser._LogParser__buffer) == 2
+
+    # 3. Push blank line (often ignored or handled by iterators, but test push)
+    line3 = "   "
+    assert parser.push(line3) is None # Blank lines are appended if buffer not empty
+    assert len(parser._LogParser__buffer) == 3
+    assert parser._LogParser__buffer[-1] == "" # After strip
+
+    # 4. Push new header - should return the *first* parsed message
+    line4 = "Origin2 Wed Apr 05 10:00:00 2017"
+    msg1 = parser.push(line4)
+    assert isinstance(msg1, LogMessage)
+    assert msg1.origin == "Origin1"
+    assert "Continuation line 1" in msg1.message
+    assert len(parser._LogParser__buffer) == 1 # Buffer now holds line4
+    assert parser._LogParser__buffer[0] == line4
+
+    # 5. Push continuation - should return None
+    line2 = "  Continuation line 2"
+    assert parser.push(line2) is None
+    assert len(parser._LogParser__buffer) == 2
+
+    # 6. Push blank line (often ignored or handled by iterators, but test push)
+    line3 = "   "
+    assert parser.push(line3) is None # Blank lines are appended if buffer not empty
+    assert len(parser._LogParser__buffer) == 3
+    assert parser._LogParser__buffer[-1] == "" # After strip
+
+    # 7. Push STOP - should return the second parsed message
+    msg2 = parser.push(STOP)
+    assert isinstance(msg2, LogMessage)
+    assert msg2.origin == "Origin2"
+    assert "Continuation line 2" in msg2.message
+    assert len(parser._LogParser__buffer) == 0 # Buffer cleared
+
+def test_08_log_message_frozen():
+    ts = datetime.now()
+    msg = LogMessage("origin", ts, Severity.INFO, 1, Facility.SYSTEM, "Test", {})
+    with pytest.raises(AttributeError):
+        msg.origin = "new_origin"
